fixing tests, splitting up geometry types into geometry_types.py
