+++ nexus_constructor/nexus_constructor_json/writer.py
@@ -9,9 +9,6 @@ import json
 from typing import List
 from nexus_constructor.data_model import (
     Component,
-    Geometry,
-    CylindricalGeometry,
-    OFFGeometry,
     PixelGrid,
     PixelMapping,
     SinglePixelId,
@@ -19,6 +16,7 @@ from nexus_constructor.data_model import (
     Translation,
     Rotation,
 )
+from nexus_constructor.geometry_types import OFFGeometry, Geometry, CylindricalGeometry
 from nexus_constructor.qml_models.instrument_model import InstrumentModel
