+++ nexus_constructor/nexus_filewriter_json/writer.py
@@ -12,12 +12,11 @@ from nexus_constructor.data_model import (
     Component,
     Translation,
     Rotation,
-    CylindricalGeometry,
-    OFFGeometry,
     PixelGrid,
     PixelMapping,
     SinglePixelId,
 )
+from nexus_constructor.geometry_types import CylindricalGeometry, OFFGeometry
 from nexus_constructor.nexus import NexusEncoder
 from nexus_constructor.qml_models.instrument_model import InstrumentModel
 from typing import List
