+++ nexus_constructor/qml_models/instrument_model.py
@@ -5,13 +5,13 @@ from nexus_constructor.data_model import (
     Component,
     Rotation,
     Translation,
-    OFFCube,
 )
 from nexus_constructor.qml_models import change_value, generate_unique_name
 from nexus_constructor.qml_models.transform_model import TransformationModel
 from nexus_constructor.off_renderer import OffMesh
 from PySide2.QtCore import Qt, QAbstractListModel, QModelIndex, Signal, Slot
 from PySide2.QtGui import QMatrix4x4, QVector3D
+from nexus_constructor.geometry_types import OFFCube
 
 
 def generate_mesh(component: Component):
